+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Add Module.php test
 - Remove unused ModuleRouteListener() instance on Module::onBootstrap as it should be already in Application's Module
 - Add Factory\Controller\Plugin\UserSettingPluginFactory test
+- Add Factory\Service\UserSettingsServiceFactory test